+++ tests/integration/tests/3396-package-indices/files/src/Lib.hs
@@ -1,8 +1,0 @@
-module Lib
-    ( someFunc
-    ) where
-
-
-someFunc :: IO ()
-someFunc = print "some func"
-